slashify kortteja
